validações de fornecedores

+++ src/FornecedoresBrasil.Business/Servicees/FornecedorServices.cs
@@ -24,7 +24,11 @@ namespace FornecedoresBrasil.Business.Servicees
                 || !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) return;
 
             //Validar se já não existe outro fornecedor com o mesmo doc.
-
+            if(_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+            {
+                Notificar("Já existe um fornecedor com este documento informado.");
+                return;
+            }
 
             await _fornecedorRepository.Adicionar(fornecedor);
         }
@@ -32,10 +36,37 @@ namespace FornecedoresBrasil.Business.Servicees
         {
             if(!ExecutarValidacao(new FornecedorValidation(), fornecedor)) return;
 
+            if(_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any())
+            {
+                Notificar("Já existe um fornecedor com este documento informado.");
+                return;
+            }
+
             await _fornecedorRepository.Atualizar(fornecedor);
         }
         public async Task Remover(Guid id)
         {
+            var fornecedor = await _fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+            if(fornecedor == null)
+            {
+                Notificar("Fornecedor não existe");
+                return;
+            }
+
+            if(fornecedor.produtos.Any())
+            {
+                Notificar("O fornecedor possui produtos cadastrados!");
+                return;
+            }
+
+            var endereco = await _fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+            if(endereco != null)
+            {
+                await _fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+            }
+
             await _fornecedorRepository.Remover(id);
         }
         public void Dispose()
